third_party/WebKit/LayoutTests/platform/mac/fast/table/table-all-rowspans-height-distribution-in-rows-except-overlapped-expected.txt

Test for chromium bug : 249600. Extra logical height is not properly spread over the rows in a row-spanning cell.
Rows in rowspan should get proportional height.
Test 1 - Three rowSpan cells
row0 col0
PASS
row1 col0 - rowspan=4	row1 col1
PASS
row2 col1
PASS
row3 col1
PASS
row4 col1
PASS
row5 col0
PASS
row6 col0
PASS
row7 col0	row7 col1 - rowspan=4
FAIL:
Expected 57 for height, but got 32. 

<tr data-expected-height="57">
	  <td>row7 col0</td>
	  <td rowspan="4" style="height:300px">row7 col1 - rowspan=4</td>
	</tr>
row8 col0
FAIL:
Expected 57 for height, but got 58. 

<tr data-expected-height="57">
      <td>row8 col0</td>
    </tr>
row9 col0
PASS
row10 col0
PASS
row11 col0	row11 col1
PASS
row12 col0
PASS
row13 col0	row13 col1 - rowspan=4
FAIL:
Expected 59 for height, but got 60. 

<tr data-expected-height="59">
	  <td>row13 col0</td>
	  <td rowspan="4" style="height:300px">row13 col1 - rowspan=4</td>
	</tr>
row14 col0
PASS
row15 col0
PASS
row16 col0
PASS
row17 col0	row17 col1
PASS
Test 2 - Three rowSpan cell and specified table width
row0 col0 - rowspan=5	row0 col1
PASS
row1 col1
PASS
row2 col1
PASS
row3 col1
PASS
row4 col1
PASS
row5 col0
PASS
row6 col0	row6 col1 - rowspan=4
FAIL:
Expected 57 for height, but got 32. 

<tr data-expected-height="57">
	  <td>row6 col0</td>
	  <td rowspan="4" style="height:300px">row6 col1 - rowspan=4</td>
	</tr>
row7 col0
FAIL:
Expected 57 for height, but got 58. 

<tr data-expected-height="57">
      <td>row7 col0</td>
    </tr>
row8 col0
PASS
row9 col0
PASS
row10 col0	row10 col1
FAIL:
Expected 19 for height, but got 34. 

<tr data-expected-height="19">
      <td>row10 col0</td>
      <td>row10 col1</td>
    </tr>
row11 col0
PASS
row12 col0	row12 col1 - rowspan=4
FAIL:
Expected 59 for height, but got 60. 

<tr data-expected-height="59">
	  <td>row12 col0</td>
	  <td rowspan="4" style="height:300px">row12 col1 - rowspan=4</td>
	</tr>
row13 col0
PASS
row14 col0
PASS
row15 col0
PASS
row16 col0	row16 col1
PASS
Test 3 - Continuous 3 rowSpan cells
row0 col0
PASS
row1 col0 - rowspan=4	row1 col1
PASS
row2 col1
PASS
row3 col1
PASS
row4 col1
PASS
row5 col0	row5 col1 - rowspan=4
FAIL:
Expected 57 for height, but got 32. 

<tr data-expected-height="57">
	  <td>row5 col0</td>
	  <td rowspan="4" style="height:300px">row5 col1 - rowspan=4</td>
	</tr>
row6 col0
FAIL:
Expected 57 for height, but got 58. 

<tr data-expected-height="57">
      <td>row6 col0</td>
    </tr>
row7 col0
PASS
row8 col0
PASS
row9 col0	row9 col1 - rowspan=4
FAIL:
Expected 59 for height, but got 60. 

<tr data-expected-height="59">
	  <td>row9 col0</td>
	  <td rowspan="4" style="height:300px">row9 col1 - rowspan=4</td>
	</tr>
row10 col0
PASS
row11 col0
PASS
row12 col0
PASS
row13 col0	row13 col1
PASS
Test 4 - Two rowSpan cells, 2 rows in first spanning cell have percent height and 2 rows in second spanning cell have fixed height
row0 col0
PASS
row1 col0 - rowspan=4	row1 col1
FAIL:
Expected 21 for height, but got 19. 

<tr data-expected-height="21">
	  <td rowspan="4" style="height:300px">row1 col0 - rowspan=4</td>
	  <td>row1 col1</td>
	</tr>
row2 col1
FAIL:
Expected 199 for height, but got 219. 

<tr data-expected-height="199" style="height:20%">
      <td>row2 col1</td>
    </tr>
row3 col1
PASS
row4 col1
FAIL:
Expected 74 for height, but got 58. 

<tr data-expected-height="74" style="height:30%">
      <td>row4 col1</td>
    </tr>
row5 col0
PASS
row6 col0
PASS
row7 col0	row7 col1 - rowspan=4
FAIL:
Expected 58 for height, but got 59. 

<tr data-expected-height="58">
	  <td>row7 col0</td>
	  <td rowspan="4" style="height:300px">row7 col1 - rowspan=4</td>
	</tr>
row8 col0
PASS
row9 col0
PASS
row10 col0
PASS
row11 col0	row11 col1
PASS
row12 col0
PASS
Test 5 - Two rowSpan cells, in first spanning cell, 2 rows have percent height and 2 rows have fixed height and in second spanning cell, 1 row have fixed height, 1 row have percent height and remaining are auto.
row0 col0
PASS
row1 col0 - rowspan=4	row1 col1
PASS
row2 col1
FAIL:
Expected 152 for height, but got 153. 

<tr data-expected-height="152" style="height:20%">
      <td>row2 col1</td>
    </tr>
row3 col1
PASS
row4 col1
PASS
row5 col0
PASS
row6 col0
PASS
row7 col0	row7 col1 - rowspan=4
PASS
row8 col0
FAIL:
Expected 218 for height, but got 144. 

<tr data-expected-height="218" style="height:30%">
      <td>row8 col0</td>
    </tr>
row9 col0
FAIL:
Expected 20 for height, but got 21. 

<tr data-expected-height="20">
      <td>row9 col0</td>
    </tr>
row10 col0
PASS
row11 col0	row11 col1
PASS
row12 col0
PASS
Test 6 - RowSpan and ColSpan.
row0 col0	row0 col1 - rowspan=3 colspan=2	row0 col2
PASS
row1 col0
PASS
row2 col0
PASS
row3 col0 - colspan=3
PASS
row4 col0
PASS
Test 7 - Mix of baseline aligned and non-baseline aligned cells.
row0 col0	row0 col1 vertical-align=top	row0 col2 vertical-align=bottom
PASS
row1 col0
PASS
row2 col0 - rowspan=4 vertical-align=text-top	row2 col1 vertical-align=sub
FAIL:
Expected 77 for height, but got 32. 

<tr data-expected-height="77">
      <td rowspan="4" style="vertical-align:text-top; height:300px">row2 col0 - rowspan=4 vertical-align=text-top</td>
      <td style="vertical-align:sub">row2 col1 vertical-align=sub</td>
    </tr>
row3 col1 vertical-align=middle
FAIL:
Expected 77 for height, but got 78. 

<tr data-expected-height="77">
      <td style="vertical-align:middle">row3 col1 vertical-align=middle</td>
    </tr>
row4 col1 vertical-align=super
PASS
row5 col1 vertical-align=text-bottom
PASS
row6 col0
PASS
Test 8 - CSS Table.
row0 col0	row0 col1	row0 col2
row1 col0	row1 col1	row1 col2
row2 col0
row3 col0
row4 col0	row4 col1	row4 col2
row5 col0	row5 col1	row5 col2
row6 col0	row6 col1
row7 col0
row8 col0
row9 col0
row10 col0
Test 9 - Table Similar to CSS table with rowspan.
row0 col0	row0 col1	row0 col2
FAIL:
Expected 21 for height, but got 19. 

<tr data-expected-height="21">
        <td rowspan="5" style="height:300px">row0 col0</td>
        <td>row0 col1</td>
        <td>row0 col2</td>
    </tr>
row1 col1	row1 col2	row1 col3
FAIL:
Expected 87 for height, but got 52. 

<tr data-expected-height="87">
        <td rowspan="3" style="height:200px">row1 col1</td>
        <td rowspan="2" style="height:100px">row1 col2</td>
        <td>row1 col3</td>
    </tr>
row2 col3
FAIL:
Expected 96 for height, but got 84. 

<tr data-expected-height="96">
        <td>row2 col3</td>
    </tr>
row3 col2
FAIL:
Expected 40 for height, but got 35. 

<tr data-expected-height="40">
        <td>row3 col2</td>
    </tr>
row4 col1	row4 col2	row4 col3
FAIL:
Expected 24 for height, but got 38. 

<tr data-expected-height="24">
        <td>row4 col1</td>
        <td>row4 col2</td>
        <td>row4 col3</td>
    </tr>
row5 col0	row5 col1	row5 col2
FAIL:
Expected 19 for height, but got 34. 

<tr data-expected-height="19">
        <td>row5 col0</td>
        <td>row5 col1</td>
        <td>row5 col2</td>
    </tr>
row6 col0	row6 col1
FAIL:
Expected 74 for height, but got 76. 

<tr data-expected-height="74">
        <td rowspan="5" style="height:200px">row6 col0</td>
        <td rowspan="5" style="height:100px">row6 col1</td>
    </tr>
row7 col1
PASS
row8 col1
PASS
row9 col1
PASS
row10 col1
PASS
